+++ Statement/Conditions.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 namespace Hector\Query\Statement;
 
 use Hector\Connection\Bind\BindParamList;
+use Hector\Connection\Driver\DriverCapabilities;
 use Hector\Query\Clause\Having;
 use Hector\Query\Clause\Where;
 use Hector\Query\StatementInterface;
@@ -43,10 +44,13 @@ class Conditions implements StatementInterface
     /**
      * @inheritDoc
      */
-    public function getStatement(BindParamList $bindParams, bool $encapsulate = false): ?string
-    {
+    public function getStatement(
+        BindParamList $bindParams,
+        ?DriverCapabilities $driverCapabilities = null,
+        bool $encapsulate = false,
+    ): ?string {
         return
-            $this->where->getStatement($bindParams, $encapsulate) .
-            $this->having->getStatement($bindParams, $encapsulate);
+            $this->where->getStatement($bindParams, $driverCapabilities, $encapsulate) .
+            $this->having->getStatement($bindParams, $driverCapabilities, $encapsulate);
     }
 }